+++ lib/eth/EthRpc.js
@@ -3,6 +3,7 @@ const promptly = require('promptly');
 const ethers = require('ethers');
 const util = require('util');
 const EventEmitter = require('events');
+const chainConfig = require('./chains');
 
 const passwordPromptAsync = util.promisify(promptly.password);
 const setTimeoutAsync = util.promisify(setTimeout);
@@ -10,6 +11,7 @@ const setTimeoutAsync = util.promisify(setTimeout);
 class EthRPC {
   constructor(config) {
     this.config = config;
+    this.chain = this.config.chain;
     this.web3 = this.getWeb3(this.config);
     this.account = this.config.account;
     this.emitter = new EventEmitter();
@@ -206,7 +208,7 @@ class EthRPC {
    * Estimates the fee.
    * @param {Object} params - The parameters for estimating the fee.
    * @param {Map} params.cache - The cache of block fees.
-   * @param {number} params.defaultEstimate - The RPC estimate of the fee.
+   * @param {number} params.defaultEstimate - The default fee value.
    * @param {number} [params.percentile=25] - The percentile to use for the estimation.
    * @param {Function} params.getFees - The function to get the fees from the blocks.
    * @returns {Promise<number>} The estimated fee.
@@ -266,7 +268,8 @@ class EthRPC {
    * @returns {Promise<number>} The estimated maximum priority fee.
    */
   async estimateMaxPriorityFee({ percentile = 25 }) {
-    const defaultEstimate = parseInt(this.web3.utils.toWei('2.5', 'gwei'));
+    const minimumFee = chainConfig[this.chain] ? chainConfig[this.chain].priorityFee : 2.5;
+    const defaultEstimate = parseInt(this.web3.utils.toWei(String(minimumFee), 'gwei'));
     const getFees = (_txs) => _txs
       .filter((tx) => !!tx.maxPriorityFeePerGas)
       .map((tx) => parseInt(tx.maxPriorityFeePerGas));
