+++ lib/index.js
@@ -6,7 +6,7 @@ const RpcClasses = {
   DOGE: require('./doge/DogeRpc'),
   LTC: require('./ltc/LtcRpc'),
   LNBTC: require('./lnd/LndRpc'),
-  MATIC: require('./matic/MaticRpc'),
+  MATIC: require('./matic/MaticRpc'), // keeping for backwards compatibility 
 };
 
 const TokenClasses = {
@@ -39,9 +39,25 @@ const TokenClasses = {
 };
 
 class CryptoRpcProvider {
+
+  /**
+   * Constructor for CryptoRpcProvider class.
+   * @param {Object} config - The configuration object.
+   * @param {string} config.chain - The chain to connect to.
+   * @param {boolean} config.isEVM - Optional flag indicating if the chain is EVM compatible.
+   * @param {string} config.host - The host address for RPC connection.
+   * @param {number} config.port - The port for RPC connection.
+   * @param {string} config.rpcPort - The port for RPC connection (alternative).
+   * @param {string} config.user - The username for RPC connection.
+   * @param {string} config.rpcUser - The username for RPC connection (alternative).
+   * @param {string} config.pass - The password for RPC connection.
+   * @param {string} config.rpcPass - The password for RPC connection (alternative).
+   * @param {string} config.protocol - The protocol for RPC connection.
+   * @param {Object} config.tokens - Optional tokens configuration.
+   */
   constructor(config) {
     this.chain = config.chain;
-    if (!RpcClasses[this.chain]) {
+    if (!RpcClasses[this.chain] && !config.isEVM) {
       throw new Error('Invalid chain specified');
     }
     this.config = Object.assign({}, config, {
@@ -51,12 +67,13 @@ class CryptoRpcProvider {
       pass: config.pass || config.rpcPass,
       protocol: config.protocol
     });
+    const rpcChain = !config.isEVM ? this.chain : 'ETH';
     this.rpcs = {
-      [this.chain]: new RpcClasses[this.chain](this.config)
+      [this.chain]: new RpcClasses[rpcChain](this.config)
     };
     if (config.tokens) {
       Object.entries(config.tokens).forEach(([token, tokenConfig]) => {
-        const TokenClass = TokenClasses[this.chain][tokenConfig.type];
+        const TokenClass = TokenClasses[rpcChain][tokenConfig.type];
         const configForToken = Object.assign(tokenConfig, this.config);
         this.rpcs[token] = new TokenClass(configForToken);
       });
